requirements.txt
streamlit
vertexai>=1.46.0
google-cloud-aiplatform>=1.48.0
google-auth
Pillow
